script_1/requirements.txt
selenium
pytest
allure-pytest
